+++ packages/react-native-reanimated/__tests__/normalizeColor.test.tsx
@@ -1,4 +1,4 @@
-import { normalizeColor } from '../src/common';
+import { normalizeColor } from '../src/Colors';
 
 describe('Test `normalizeColor` function', () => {
   describe('Only compliant color are accepted', () => {
