+++ packages/react-native-reanimated/src/animation/styleAnimation.ts
@@ -1,7 +1,8 @@
 'use strict';
 import { logger } from 'react-native-worklets';
 
-import { ColorProperties, processColor } from '../common';
+import { ColorProperties } from '../Colors';
+import { processColor } from '../common';
 import type {
   AnimatableValue,
   AnimatedStyle,
