+++ packages/react-native-reanimated/src/common/processors/transformOrigin.ts
@@ -1,6 +1,6 @@
 'use strict';
 'worklet';
-import { ReanimatedError } from '..';
+import { ReanimatedError } from '../errors';
 import type {
   NormalizedTransformOrigin,
   TransformOrigin,
